+++ services/userService.js
@@ -1,24 +1,57 @@
-const Cliente = require('../models/userModel'); 
+const Cliente = require('../models/userModel');
+const facturapiService = require('../apis/facturapi');
 
+// Crear cliente en MongoDB y Facturapi
 async function crearCliente(data) {
-  const cliente = new Cliente(data);
-  await cliente.save();
-  return cliente;
+  // Primero, crear en Facturapi
+  const clienteFacturapi = await facturapiService.crearCliente(data);
+
+  // Guardar también el ID de Facturapi en Mongo
+  const clienteMongo = new Cliente({
+    ...data,
+    facturapiId: clienteFacturapi.id
+  });
+
+  await clienteMongo.save();
+  return clienteMongo;
 }
 
+// Obtener todos los clientes desde Mongo
 async function obtenerClientes() {
   return await Cliente.find();
 }
 
+// Obtener cliente por ID desde Mongo
 async function obtenerClientePorId(id) {
   return await Cliente.findById(id);
 }
 
+// Actualizar cliente en Mongo y Facturapi
 async function actualizarCliente(id, data) {
+  const clienteMongo = await Cliente.findById(id);
+  if (!clienteMongo) throw new Error('Cliente no encontrado');
+
+  // Actualizar en Facturapi
+  await facturapiService.actualizarCliente(clienteMongo.facturapiId, {
+    legal_name: data.legal_name,
+    email: data.email,
+    tax_id: data.rfc,
+    address: data.address
+  });
+
+  // Actualizar en Mongo
   return await Cliente.findByIdAndUpdate(id, data, { new: true });
 }
 
+// Eliminar cliente en Mongo y Facturapi
 async function eliminarCliente(id) {
+  const clienteMongo = await Cliente.findById(id);
+  if (!clienteMongo) throw new Error('Cliente no encontrado');
+
+  // Eliminar en Facturapi
+  await facturapiService.eliminarCliente(clienteMongo.facturapiId);
+
+  // Eliminar en Mongo
   return await Cliente.findByIdAndDelete(id);
 }
